fix formatting bug during push
The output of the progress message would not be cleared at the end, leading to garbled text when the final tag is printed.
