+++ pkg/push/pusheager.go
@@ -3,8 +3,6 @@ package push
 import (
 	"context"
 	"fmt"
-	"os"
-	"time"
 
 	"github.com/malt3/go-containerregistry/pkg/authn"
 	"github.com/malt3/go-containerregistry/pkg/name"
@@ -90,21 +88,4 @@ func (p *eagerPusher) PushIndex(ctx context.Context, reference string, req PushI
 		return "", err
 	}
 	return digest.String(), nil
-}
-
-func progressPrinter(updates <-chan registryv1.Update) {
-	var lastUpdate time.Time
-	for update := range updates {
-		relative := float64(update.Complete) / float64(update.Total) * 100
-		if update.Error != nil {
-			fmt.Fprintf(os.Stderr, "Error: %v\n", update.Error)
-			continue
-		}
-		if time.Since(lastUpdate) < 10*time.Millisecond {
-			// Avoid printing too frequently
-			continue
-		}
-		fmt.Fprintf(os.Stderr, "Progress: %.2f %% (%v / %v bytes)\r", relative, update.Complete, update.Total)
-		lastUpdate = time.Now()
-	}
 }
